[Gold V] Title: 퇴사 2, Time: 188 ms, Memory: 7880 KB -BaekjoonHub

+++ 백준/Gold/15486. 퇴사 2/퇴사 2.cc
@@ -9,7 +9,7 @@
 using namespace std;
 
 
-long long N;
+int N;
 int D[1500051];
 int T, P;
